Added new file `src/parseBar.ts`
This file contains functions and types for extracting
the data needed to generate duplicate comment bars from
a string containing the lines of a comment bar.

This will allow new features such as changing the text of
the bar without changing the width, along with changing
the width while keeping the text centered.